feat(adieu-razzle): moved common build function into webpack tools
